requirements.txt
python-dotenv
requests
spotipy
python-telegram-bot
